Add options for dparser & extract build.
- Added echoing option and garbage collection options
- Change the build to be outside of dparser to break rex dependency.

+++ R/dparser.R
@@ -38,312 +38,29 @@
 ##' }
 ##'
 ##' The result is natural grammars and powerful parsing.
+##'
+##' The R based tree parsing in \code{\link{dparse}} creates dlls on
+##' the fly based on C code. By default the echoing of the compile is
+##' disabled, but you can change this by
+##' \code{options(dparser.echo.compile=FALSE)}
+##'
+##' @template garbage
+##'
 ##' @template grammar
 ##' @useDynLib dparser cDparser dparse_dparser_gram
 ##' @importFrom methods new
 "_PACKAGE"
+.onLoad <- function(libname, pkgname){ ## nocov start
+    op.dparser <- list(dpaser.rm.named.parser.dll=FALSE,
+                       dpaser.rm.unnamed.parser.dll=TRUE,
+                       dparser.echo.compile=FALSE)
+    op <- options();
+    w <- !(names(op.dparser) %in% names(op))
+    if (any(w)) options(op.dparser[w]);
+} ## nocov end
 
-rex::register_shortcuts("dparser");
 refresh <- function(){ # nocov start
-    if (!file.exists(devtools::package_file("src/dparser"))){
-        owd <- getwd();
-        setwd(devtools::package_file("src"));
-        system("git clone https://github.com/jplevyak/dparser")
-        setwd(owd);
-    }
-    missingFns <- c("finalize_productions", # acutally missing
-                    "d_warn",
-                    "d_fail",
-                    "stack_push_internal");
-    globalIntVars <- c("d_use_r_headers",
-                       "d_rdebug_grammar_level",
-                       "d_use_file_name",
-                       "d_verbose_level",
-                       "d_debug_level");
-    globalCharVars <- c("d_file_name");
-    if (file.exists(devtools::package_file("src/dparser"))){
-        owd <- getwd();
-        setwd(devtools::package_file("src/dparser"));
-        build <- gsub("([^ ]*) .*","\\1",system("git show-ref heads/master",intern=TRUE));
-        setwd(owd);
-        cat(sprintf("dparser build %s\n", build));
-        for (f in c("dparse.h", "dparse_tables.h", "dsymtab.h", "gram.h", "gramgram.h", "lex.h",
-                    "lr.h", "mkdparse.h", "parse.h", "read_binary.h", "scan.h", "util.h", "write_tables.h",
-                    "gram.c", "grammar.g.c", "lex.c", "lex.h", "lr.c", "mkdparse.c", "scan.c", "symtab.c",
-                    "util.c", "write_tables.c", "d.h", "read_binary.c", "parse.c", "version.c")){
-            cat(sprintf("\tf: %s\n", f));
-            unlink(devtools::package_file("src/", f));
-            d <- readLines(devtools::package_file("src/dparser/", f));
-            if (any(f == c("gram.h", "lex.h", "lr.h", "parse.h", "read_binary.h", "scan.h", "util.h", "write_tables.h"))){
-                ## Fix headers to work with C++
-                d[1] <- sprintf('#if defined(__cplusplus)\nextern "C" {\n#endif\n%s', d[1]);
-                d[length(d)] <- sprintf('%s\n#if defined(__cplusplus)\n}\n#endif\n', d[length(d)]);
-                if (f == "gram.h"){
-                    ## Fix add_pass_code to be on one line.
-                    w <- which(regexpr("void add_pass_code", d) != -1);
-                    d[w] <- sprintf("%s %s", d[w], gsub("^[ \t]*", "", d[w + 1]));
-                    d <- d[-(w + 1)];
-                }
-            }
-            if (f == "d.h"){
-                w <- which(regexpr('#define (REALLOC|MALLOC|FREE|CALLOC)', d) != -1);
-                d <- d[-w];
-                w <- which(regexpr('#include "arg.h"', d) != -1);
-                d <- d[-w];
-                w <- which(regexpr('#include <string.h>', d) != -1)
-                ver  <- readLines(devtools::package_file("src/dparser/Makefile"));
-                major <- gsub("^ *MAJOR *= *", "", ver[which(regexpr("^ *MAJOR *=", ver) != -1)]);
-                minor <- gsub("^ *MINOR *= *", "", ver[which(regexpr("^ *MINOR *=", ver) != -1)]);
-                d <- c(d[1:w], "#include <R.h>", "#include <Rinternals.h>",
-                       sprintf("#define D_MAJOR_VERSION %s", major),
-                       sprintf("#define D_MINOR_VERSION %s", minor),
-                       sprintf('#define D_BUILD_VERSION "R-%s"', build),
-                       d[seq(1 + w, length(d))]);
-            }
-            d <- gsub("REALLOC", "R_chk_realloc", d);
-            d <- gsub("MALLOC[(]", "R_chk_calloc(1,", d);
-            d <- gsub("CALLOC[(]", "R_chk_calloc(1,", d);
-            d <- gsub("FREE", "Free", d);
-            d <- gsub("([ \t])printf[(]", "\\1Rprintf(", d);
-            if (f == "parse.c"){
-                w <- which(regexpr("^ *syntax_error_report_fn", d) != -1);
-                if (regexpr("^[ \t]*$", d[w - 2]) != -1){
-                    d[w - 2] <- "\nchar * d_file_name;\n int  d_use_file_name = 0;\n"
-                    while(regexpr("^[ \t]*ZNode [*]z", d[w]) == -1){
-                        w <- w + 1;
-                    }
-                    d[w] <- sprintf("%s\n  if (d_use_file_name){\n    fn = d_dup_pathname_str(d_file_name);\n }\n", d[w]);
-                } else {
-                    stop("something is wrong.")
-                }
-                d <- gsub("fprintf[(]stderr[ \t]*,", "Rprintf(", d);
-            }
-            if (f == "write_tables.c"){
-                w  <- which(regexpr('#include "dparse_tables.h"', d, fixed=TRUE) != -1);
-                d <- c(d[1:w], "int d_use_r_headers = 0;", d[seq(w + 1, length(d))]);
-                w <- which(regexpr("Available at http://dparser.sf.net", d, fixed=TRUE) != -1);
-                for (i in rev(w)){
-                    fp <- gsub(" *fprintf[(] *([^ ]*) *,.*", "\\1", d[i]);
-                    d[i] = sprintf('fprintf(%s, "\t/*\\n  Generated by R\\\'s mkdparse a port of Make DParser Version %%s\\n", ver);
-\tfprintf(%s,"  R available at https://github.com/mattfidler/dparser\\n");
-\tfprintf(%s,"  Original dparser Available at http://dparser.sf.net\\n*/\\n\\n\\n");
-\tif (d_use_r_headers){
-\t\tfprintf(%s,"\\n\\n#include <R.h>\\n#include <Rinternals.h>\\n#define printf Rprintf\\n\\n");
-\t}',fp, fp, fp, fp);
-                    d <- d[-(i - 1)];
-                }
-            }
-            if (f == "util.c"){
-                w <- which(regexpr("^ *d_warn *[(]", d) != -1);
-                w2 <- w + 1;
-                while (regexpr("}", d[w2]) == -1){
-                    w2 <- w2 + 1;
-                }
-                d  <- c(d[seq(1, w - 1)],
-                        'd_warn(const char *str, ...) {
-  char nstr[256];
-  char outstr[256*2];
-  va_list ap;
-  va_start(ap, str);
-  snprintf(nstr, 255, "%s", str);
-  vsprintf(outstr, nstr, ap);
-  va_end(ap);
-  warning(outstr);
-}',
-d[seq(w2 + 1, length(d))]);
-                w <- which(regexpr("^ *d_fail *[(] *const +char", d) != -1);
-                w2 <- w + 1;
-                while (regexpr("}", d[w2]) == -1){
-                    w2 <- w2 + 1;
-                }
-                d  <- c(d[seq(1, w - 1)],
-                        'd_fail(const char *str, ...) {
-  char nstr[256];
-  char outstr[256*2];
-  va_list ap;
-  va_start(ap, str);
-  snprintf(nstr, 255, "Parser Fail: %s", str);
-  vsprintf(outstr, nstr, ap);
-  va_end(ap);
-  error(outstr);
-}',
-d[seq(w2 + 1, length(d))]);
-            }
-
-            sink(devtools::package_file("src/", f));
-            cat(paste(d, collapse="\n"));
-            sink()
-        }
-    }
-    id <- rex::rex(one_of("_", "a":"z", "A":"Z"), any_of("_", "a":"z", "A":"Z", "0":"9"));
-    fnRex <- rex::rex(start, any_spaces, capture(id), spaces,
-                      capture(any_of("*")), any_spaces, capture(id), any_spaces,
-                      one_of("("), capture(except_any_of(")")), one_of(")"), any_spaces, one_of(";"));
-    commentRex <- rex::rex(or(group("/*", anything, "*/"), group("//", anything, end)));
-    comSep <- rex::rex(any_spaces, ",", any_spaces);
-    argRex <- rex::rex(start, any_spaces, capture(except_some_of("*")), spaces, capture(any_of("*")),
-                       any_spaces, capture(id));
-    fns <- c();
-    calls <- c();
-    dparser <- "SEXP dparse_sexp(SEXP sexp_fileName, SEXP sexp_start_state, SEXP sexp_save_parse_tree, SEXP sexp_partial_parses, SEXP sexp_compare_stacks, SEXP sexp_commit_actions_interval, SEXP sexp_fixup, SEXP sexp_fixup_ebnf, SEXP sexp_nogreedy, SEXP sexp_noheight, SEXP sexp_use_filename, SEXP sexp_sizeof_parse_node, SEXP sexp_verbose, SEXP fn, SEXP env, D_ParserTables pt);\n"
-    headers <- c("dparse.h", "dparse_tables.h", "dsymtab.h", "gram.h", "gramgram.h", "lex.h",
-                 "lr.h", "mkdparse.h", "parse.h", "read_binary.h", "scan.h", "util.h", "write_tables.h");
-    defs <- c();
-    for (f in c("dparse.h", "dparse_tables.h", "dsymtab.h", "gram.h", "gramgram.h", "lex.h",
-                "lr.h", "mkdparse.h", "parse.h", "read_binary.h", "scan.h", "util.h", "write_tables.h", "")){
-        if (f == ""){
-            txt <- c(sprintf("void set_%s(int x);", globalIntVars),
-                     sprintf("int get_%s();", globalIntVars),
-                     sprintf("void set_%s(char *x);", globalCharVars),
-                     dparser);
-        } else {
-            txt <- suppressWarnings({readLines(devtools::package_file(sprintf("src/%s", f)))});
-        }
-        ## print(txt);
-        txt <- gsub(commentRex, "", txt[regexpr(fnRex, txt, perl=TRUE) != -1]);
-        for (txti in txt){
-            fnType <- gsub(fnRex, "\\1", txti, perl=TRUE);
-            stars <- gsub(fnRex, "\\2", txti, perl=TRUE);
-            fnName <- gsub("\n", "", gsub(fnRex, "\\3", txti, perl=TRUE));
-            if (fnName == "d_free"){
-                fnArg <- "void *x";
-            } else {
-                fnArg <- gsub(fnRex, "\\4", txti, perl=TRUE);
-            }
-            if (!any(fnName == missingFns)){
-                if (f != ""){
-                    defs <- c(defs, txti);
-                }
-                argsTot <- strsplit(fnArg, comSep)[[1]];
-                args <- gsub(argRex, "\\1\\2", argsTot, perl=TRUE);
-                arge <- gsub(argRex, "\\3", argsTot, perl=TRUE);
-                fn <- sprintf("%s %s%s(%s){\n  static %s %s(*fun)(%s)=NULL;\n  if (fun == NULL) fun = (%s%s (*)(%s)) R_GetCCallable(\"dparser\",\"%s\");\n  return fun(%s);\n}\n",
-                              fnType, stars, fnName, fnArg,
-                              fnType, stars, paste(args, collapse=", "),
-                              fnType, stars, paste(args, collapse=", "),
-                              fnName, paste(arge, collapse=", "));
-                fns <- c(fns, fn);
-                call <- sprintf("  R_RegisterCCallable(\"dparser\",\"%s\",(DL_FUNC) %s);\n", fnName, fnName);
-                calls <- c(call, calls);
-            }
-        }
-    }
-    fns <- sprintf("/*
-Header file for using internal C-level dparser functions in dparser-R (generated).
-*/
-#ifndef __dparser_H__
-#define __dparser_H__
-#include <R.h>
-#include <Rinternals.h>
-#include <Rdefines.h>
-#include <Rconfig.h>
-#include <R_ext/Rdynload.h>
-#include \"gramgram.h\"
-#include \"d.h\"
-#include \"mkdparse.h\"
-#include \"dparse.h\"
-#include \"read_binary.h\"
-#if defined(__cplusplus)
-extern \"C\" {
-#endif
-%s
-void
-d_fail(const char *str, ...) {
-  char nstr[256];
-  char outstr[256*2];
-  va_list ap;
-  va_start(ap, str);
-  snprintf(nstr, 255, \"Parser Fail: %%s\", str);
-  vsprintf(outstr, nstr, ap);
-  va_end(ap);
-  error(outstr);
-}
-
-void
-d_warn(const char *str, ...) {
-  char nstr[256];
-  char outstr[256*2];
-  va_list ap;
-  va_start(ap, str);
-  snprintf(nstr, 255, \"%%s\", str);
-  vsprintf(outstr, nstr, ap);
-  va_end(ap);
-  warning(outstr);
-}
-
-#if defined(__cplusplus)
-}
-#endif
-#endif\n", paste(fns, collapse="\n"));
-    sink(devtools::package_file("src/dparser.h"));
-    cat(fns);
-    sink();
-    cat(sprintf("\tf: dparser.h\n"));
-    sink(devtools::package_file("src/dparser.c"));
-    cat(sprintf("/*
-Register C callables to R.
-*/
-#include <R.h>
-#include <Rinternals.h>
-#include <Rdefines.h>
-#include <Rconfig.h>
-#include <R_ext/Rdynload.h>
-#include \"gramgram.h\"
-#include \"d.h\"
-#include \"mkdparse.h\"
-#include \"dparse.h\"
-%s
-%s
-%s
-%s
-
-void R_init_dparser(DllInfo *info){
-%s}
-",dparser, paste(sprintf("extern int %s;\nvoid set_%s(int x){\n  %s = x;\n}\nint get_%s(){\n return %s;\n}\n", globalIntVars, globalIntVars, globalIntVars, globalIntVars, globalIntVars), collapse="\n"),
-paste(sprintf("extern char * %s;\nvoid set_%s(char *x){\n  %s=x;\n}\n", globalCharVars, globalCharVars, globalCharVars), collapse="\n"),
-paste(defs, collapse="\n"),paste(calls, collapse="")));
-    sink();
-    cat(sprintf("\tf: dparser.c\n"))
-    dtest <- devtools::package_file("src/dparser/tests");
-    ttest <- devtools::package_file("tests/testthat");
-    for (f in list.files(dtest,"[.]g", full.names=TRUE)){
-        cat(sprintf("\tf: %s\n", f));
-        file.copy(f, ttest, TRUE);
-    }
-    sink(devtools::package_file("R/version.R"))
-    cat("##\' Version and repository for this dparser package.
-##\'
-##\' @return A character vector with the version and repository.
-##\' @author Matthew L. Fidler
-##\' @keywords internal
-##\' @export
-dpVersion <- function(){return(c(version=\"");
-    cat(gsub("Version: +", "", readLines(devtools::package_file("DESCRIPTION"), 2)[2]))
-    cat("\",repo=\"");
-    cat("https://github.com/")
-    tmp <- readLines(devtools::package_file(".git/config"))
-    cat(gsub("\\.git$", "", gsub(".*git@github.com:", "", tmp[which(tmp == '[remote "origin"]')[1]+1])))
-    cat("\"))}\n");
-    sink();
-    sink(devtools::package_file("R/Rparse.R"))
-    tmp <- gsub("\"", "\\\\\"", readLines(devtools::package_file("build/Rparse.c")));
-    cat("##\' Rcpp code for R parser
-##\'
-##\' @return A string for generating Rcpp tree walker
-##\' @author Matthew L. Fidler
-##\' @keywords internal
-##\' @export
-dpRparse <- function(){return(\"")
-    cat(paste(tmp, collapse="\\n"));
-    cat("\\n\");}\n");
-    sink();
-    cat("Update README\n");
-    owd <- getwd();
-    on.exit({setwd(owd)});
-    setwd(devtools::package_file());
-    knitr::knit(devtools::package_file("README.Rmd"))
-    devtools::document();
-    devtools::load_all();
+    source(devtools::package_file("build/refresh.R"))
 } # nocov end
 
 ##' Make a dparser c file based on a grammer
@@ -546,18 +263,22 @@ dpIncludeDir <- function(...){
 ##'
 ##' @param fileext file extension of temporary file to create
 ##'
-##' @return a list of two elements:
+##' @param envir Environment to deparse variables
+##'
+##' @return a list of three elements:
 ##' \describe{
 ##'
 ##' \item{\code{file}}{The file name of etiher the tempoary file or the real file}
 ##'
-##' \item{code{use_file_name}}{If the file name was used (\code{TRUE}), or a temporary file was created(\code{FALSE})}
+##' \item{\code{use_file_name}}{If the file name was used (\code{TRUE}), or a temporary file was created(\code{FALSE})}
+##'
+##' \item{\code{md5}}{md5 of the model}
 ##'
 ##' }
 ##' @author Matthew L. Fidler
 ##' @keywords internal
 ##' @export
-dpGetFile <- function(file, fileext=""){
+dpGetFile <- function(file, fileext="", envir=parent.frame(1)){
     ret <- list(use_file_name=FALSE);
     if (class(substitute(file)) == "call"){
         file <- file;
@@ -571,16 +292,20 @@ dpGetFile <- function(file, fileext=""){
         if (file.exists(file)){
             ret$use_file_name <- TRUE;
         }
+    } else if (class(file) == "name"){
+        file <- eval(file, envir=envir);
     } else {
         stop(sprintf("Cant figure out how to handle the file argument (%s).", class(file)));
     }
     if (!ret$use_file_name){
+        ret$md5 <- digest::digest(file);
         tmpf <- tempfile(fileext=fileext);
         sink(tmpf);
         cat(file);
         sink();
         ret$file <- tmpf;
     } else {
+        ret$md5 <- digest::digest(file, file=TRUE);
         ret$file <- file;
     }
     return(ret);
@@ -624,6 +349,8 @@ gc.items <- c()
 ##' This will delete the dynamically created dll upon garbage
 ##' collection.
 ##'
+##' @template garbage
+##'
 ##' @param env Environment that is being garbage collected.
 ##' @return Nothing.
 ##' @author Matthew L. Fidler
@@ -638,8 +365,12 @@ gc.dparser <- function(env){
             assignInMyNamespace("gc.items", tmp);
         } else {
             if (file.exists(dll.file)){
-                rc <- try(dyn.unload(dll.file), silent = TRUE);
-                unlink(dll.file)
+                generated <- env$generated;
+                if ((generated && getOption("dpaser.rm.unnamed.parser.dll", TRUE)) ||
+                    (!generated && getOption("dpaser.rm.named.parser.dll", FALSE))){
+                    rc <- try(dyn.unload(dll.file), silent = TRUE);
+                    unlink(dll.file)
+                }
             }
             tmp <- gc.items;
             tmp <- tmp[names(tmp) != dll.file];
@@ -675,6 +406,9 @@ gc.dparser <- function(env){
 ##'     \code{\link{dparse_gram}}.  This function would be able to
 ##'     parse arbitary grammars the way you may want with your own
 ##'     user supplied function.
+##'
+##' @template garbage
+##'
 ##' @seealso \code{\link{dparse_gram}}, \code{\link{mkdparse}}
 ##' @export
 dparse <- function(grammar,
@@ -690,7 +424,7 @@ dparse <- function(grammar,
                    use_file_name=TRUE,
                    parse_size=1024,
                    verbose_level=0, ...){
-    lst <- dpGetFile(substitute(grammar), ".g");
+    lst <- dpGetFile(substitute(grammar), ".g", envir=parent.frame(1));
     grammar <- lst$file;
     if (missing(use_file_name)){
         use_file_name <- as.integer(lst$use_file_name);
@@ -708,8 +442,12 @@ dparse <- function(grammar,
     fixup_ebnf <- as.integer(fixup_ebnf);
     nogreedy <- as.integer(nogreedy);
     noheight <- as.integer(noheight);
-    gram <- gsub("[.]g", "", basename(grammar))
-    md5 <- digest::digest(grammar, file=TRUE);
+    md5 <- lst$md5;
+    if (lst$use_file_name){
+        gram <- gsub("[.]g", "", basename(grammar))
+    } else {
+        gram <- sprintf("dp_%s", md5);
+    }
     pkg <- sprintf("%s_%s", gram, .Platform$r_arch)
     dll.file <- sprintf("%s_%s%s", gram, .Platform$r_arch, .Platform$dynlib.ext);
     if (file.exists(dll.file)){
@@ -744,7 +482,8 @@ dparse <- function(grammar,
         cmd <- sprintf("%s/bin/R CMD SHLIB %s_%s.c",
                        Sys.getenv("R_HOME"), gram, .Platform$r_arch);
         sh <- "system";
-        do.call(sh,list(cmd,ignore.stdout=FALSE,ignore.stderr=FALSE));
+        do.call(sh,list(cmd,ignore.stdout=!getOption("dparser.echo.compile", TRUE),
+                        ignore.stderr=!getOption("dparser.echo.compile", TRUE)));
         setwd(owd);
         file.copy(file.path(tmpd, dll.file), getwd());
         unlink(tmpd, recursive=TRUE);
@@ -770,7 +509,7 @@ dparse <- function(grammar,
     }));
     sym <- getNativeSymbolInfo(sprintf('dparse_%s', gram), PACKAGE=pkg);
     body <- bquote({
-        lst <- dpGetFile(substitute(file));
+        lst <- dpGetFile(substitute(file), envir=parent.frame(1));
         file <- lst$file;
         start_state         <- as.integer(start_state);
         save_parse_tree     <- as.integer(save_parse_tree);
@@ -810,6 +549,7 @@ dparse <- function(grammar,
     body(fun) <- body;
     ret <- new("dparserFunction", env=new.env());
     ret@.Data <- fun;
+    ret@env$generated <- (!lst$use_file_name);
     ret@env$grammar <- normalizePath(grammar)
     ret@env$dll.file <- normalizePath(dll.file)
     ## Cleanup
@@ -869,7 +609,7 @@ dparse_gram <- function(file,
                         use_file_name=TRUE,
                         parse_size=1024,
                         verbose_level=0){
-    lst <- dpGetFile(file);
+    lst <- dpGetFile(file, envir=parent.frame(1));
     file <- lst$file;
     start_state         <- as.integer(start_state);
     save_parse_tree     <- as.integer(save_parse_tree);
